backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.30.*
sqlalchemy[asyncio]==2.0.*
asyncpg==0.30.*
geoalchemy2==0.18.*
shapely==2.*
alembic==1.13.*
celery==5.4.*
redis==5.*
pydantic-settings==2.*
pyjwt==2.12.*
pwdlib[argon2]==0.3.*
python-multipart
pytest==8.*
pytest-asyncio==0.24.*
httpx==0.27.*
anyio==4.*
